palivla_digit/pyproject.toml
[project]
name = "big-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "big-vision",
    "chex>=0.1.86",
    "distrax>=0.1.5",
    "dlimp",
    "einops>=0.8.0",
    "flax>=0.9.0",
    "ipykernel",
    "jax[cuda12]>=0.4.33",
    "matplotlib>=3.9.2",
    "ml-collections>=0.1.1",
    "numpy<2.0.0",
    "octo",
    "optax>=0.2.3",
    "orbax-checkpoint>=0.7.0",
    "orbax>=0.1.9",
    "overrides>=7.7.0",
    "pip",
    "scalax>=0.2.4",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "sentencepiece>=0.2.0",
    "tensorflow-datasets>=4.9.6",
    "tensorflow-text-nightly>=2.19.0.dev20241014",
    "tf-nightly>=2.19.0.dev20241014",
    "wandb>=0.18.3",
]

[tool.uv.sources]
dlimp = { path = "dlimp", editable = true }
octo = { path = "orca", editable = true }
big-vision = { workspace = true }
